Updated Calculations.java
added more exception handing,
when dividing by zero, the system will prompt for a number not equal to zero and will finish the calculation with the new number instead of shutting down. If a user did not enter a valid number on the second try, Illegal state exception,and no such element exception is caught,which will bring you back to the menu and ask you to restart instead of shutting down.Did the same to modulus and factorial,Made sure factorials over than 170, give errors as 171! and more are far too large to computer